fix: support read list of fixed-size-list (#93)
In our scenario, there will be cases of a list of fixed-size lists,
which is written using Ray. When reading it with Spark, an error occurs,
but after making modifications, it can be read normally

Using Spark to define and write a nested fixed-size list is a bit
complicated, and it's currently not easy to write tests. 🤔