feat: support dataframe load path API (#131)
The current lance read and write requires catalog configuration, and it
does not support s3://path/to/table.lance for direct read and write
paths. The purpose of this PR is to not only support catalog mode, but
also support dataframe load API to read and write lance.

example:
```
spark.read.format("lance").load("s3://path/to/a.lance")
```